remove osrm text instructions
