+++ Cartfile.resolved
@@ -1,8 +1,7 @@
 binary "https://www.mapbox.com/ios-sdk/Mapbox-iOS-SDK.json" "3.6.4"
 github "1ec5/Solar" "5ce2ca7baaa82b1abb0467642ae237faf674b645"
-github "Project-OSRM/osrm-text-instructions.swift" "v0.5.0"
+github "aws/aws-sdk-ios" "2.6.5"
 github "facebook/ios-snapshot-test-case" "2.1.4"
-github "frederoni/aws-sdk-ios" "2.6.5"
 github "mapbox/MapboxDirections.swift" "v0.12.1"
 github "mapbox/mapbox-telemetry-ios" "v0.2.11"
 github "mapbox/turf-swift" "v0.0.4"
