+++ MapboxCoreNavigation.podspec
@@ -41,7 +41,6 @@ Pod::Spec.new do |s|
   s.module_name = "MapboxCoreNavigation"
 
   s.dependency "MapboxDirections.swift", "~> 0.12"
-  s.dependency "OSRMTextInstructions", "~> 0.5"
   s.dependency "MapboxMobileEvents", "~> 0.2"
   s.dependency "Turf", "~> 0.0.4"
