+++ MapboxNavigation-Documentation.podspec
@@ -45,7 +45,6 @@ Pod::Spec.new do |s|
 
   s.dependency "MapboxDirections.swift", "~> 0.12"
   s.dependency "Mapbox-iOS-SDK", "~> 3.6"
-  s.dependency "OSRMTextInstructions", "~> 0.5"
   s.dependency "SDWebImage", "~> 4.1"
   s.dependency "AWSPolly", "~> 2.6"
   s.dependency "MapboxMobileEvents", "~> 0.2"
